Copying attachment file to internal app sandbox, debugging multipart upload

+++ app/src/main/java/com/mercata/pingworks/models/MessageFilePartInfo.kt
@@ -3,9 +3,9 @@ package com.mercata.pingworks.models
 data class MessageFilePartInfo(
     val urlInfo: URLInfo,
     val messageId: String,
-    val part: Long,
+    val part: Int,
     val size: Long,
     val checksum: String? = null,
     val offset: Long? = null,
-    val totalParts: Long,
+    val totalParts: Int,
 )
